Moves logic to editor view
