+++ lib/assets/javascripts/cartodb3/editor/map/map-view.js
@@ -1,13 +1,10 @@
 var $ = require('jquery');
 var Backbone = require('backbone');
 var cdb = require('cartodb-deep-insights.js');
-var EditorWidgetsView = require('../widgets/widgets-view');
-var createTextLabelsTabPane = require('../../components/tab-pane/create-text-labels-tab-pane');
-var StackLayoutView = require('../../components/stack-layout/stack-layout-view');
-var WidgetsFormContentView = require('../widgets/widgets-form/widgets-form-content-view');
+var EditorView = require('../editor-view');
 var Template = require('./template.tpl');
-var TabPaneTemplate = require('./tab-pane-template.tpl');
 var AddWidgetsView = require('../add-widgets/add-widgets-view');
+var StackLayoutView = require('../../components/stack-layout/stack-layout-view');
 
 module.exports = cdb.core.View.extend({
   initialize: function (opts) {
@@ -24,70 +21,30 @@ module.exports = cdb.core.View.extend({
   },
 
   render: function () {
+    var self = this;
     this.$el.html(this.template({
       vis_name: this._visDefinitionModel.get('name')
     }));
 
-    var self = this;
-    // TODO for now we only render widgets, but at some point this should be "wrapped" by a EditorView
     var stackViewCollection = new Backbone.Collection([{
       createStackView: function (stackLayoutModel, opts) {
-        return new EditorWidgetsView({
+        return new EditorView({
           layerDefinitionsCollection: self._layerDefinitionsCollection,
           widgetDefinitionsCollection: self._widgetDefinitionsCollection,
           stackLayoutModel: stackLayoutModel
         });
       }
-    }, {
-      createStackView: function (stackLayoutModel, opts) {
-        var widgetDefinitionModel = opts[0];
-        return new WidgetsFormContentView({
-          widgetDefinitionModel: widgetDefinitionModel,
-          layerDefinitionsCollection: self._layerDefinitionsCollection,
-          stackLayoutModel: stackLayoutModel
-        });
-      }
     }]);
 
-    var mapTabPaneView = createTextLabelsTabPane([{
-      label: _t('editor.layers.title'),
-      selected: true,
-      createContentView: function () {
-        return new cdb.core.View();
-      }
-    }, {
-      label: _t('editor.elements.title'),
-      createContentView: function () {
-        return new cdb.core.View();
-      }
-    }, {
-      label: _t('editor.widgets.title'),
-      createContentView: function () {
-        return new StackLayoutView({
-          collection: stackViewCollection
-        });
-      }
-    }], {
-      tabPaneOptions: {
-        tagName: 'nav',
-        className: 'CDB-NavMenu',
-        template: TabPaneTemplate,
-        tabPaneItemOptions: {
-          tagName: 'li',
-          className: 'CDB-NavMenu-Item'
-        }
-      },
-      tabPaneItemLabelOptions: {
-        tagName: 'button',
-        className: 'CDB-NavMenu-Link'
-      }
+    var stackLayoutView = new StackLayoutView({
+      collection: stackViewCollection
     });
 
-    this.$('.js-content').append(mapTabPaneView.render().$el);
+    this.$('.js-content').append(stackLayoutView.render().$el);
 
     // TODO tmp; move to edit-content-view once the top-level tab pane is implemented
     var $addWidgetsBtn = $('<button class="CDB-Button CDB-Button--primary"><span>Add widgets</span></button>');
-    mapTabPaneView.$el.append($addWidgetsBtn);
+    stackLayoutView.$el.append($addWidgetsBtn);
     $addWidgetsBtn.on('click', function () {
       // Open a add-widgets-modal on page load
       self._modals.create(function (modalModel) {
