+++ lib/assets/javascripts/cartodb3/editor/tab-pane-template.tpl
