+++ lib/assets/javascripts/cartodb3/editor/widgets/widgets-form/widgets-form-content-view.js
@@ -3,7 +3,7 @@ var $ = require('jquery');
 var WidgetsFormContentDataView = require('./widgets-form-content-data-view');
 var WidgetsFormContentStyleView = require('./widgets-form-content-style-view');
 var createTextLabelsTabPane = require('../../../components/tab-pane/create-text-labels-tab-pane');
-var TabPaneTemplate = require('../../map/tab-pane-template.tpl');
+var TabPaneTemplate = require('../../tab-pane-template.tpl');
 
 /**
  * View to render all necessary for the widget form
